city build in util
