+++ src/main/java/com/canoo/cog/ui/city/util/StageUtil.java
@@ -1,7 +1,11 @@
 package com.canoo.cog.ui.city.util;
 
-import com.canoo.cog.ui.city.CityBuilder;
+import java.util.List;
+
+import com.canoo.cog.solver.CityNode;
+import com.canoo.cog.ui.city.model.Building;
 import com.canoo.cog.ui.city.model.City;
+import com.canoo.cog.ui.city.model.Hood;
 import com.canoo.cog.ui.city.model.style.CityStyle;
 import com.canoo.cog.ui.city.model.text.Info;
 import javafx.beans.property.SimpleStringProperty;
@@ -70,5 +74,45 @@ public class StageUtil {
 
     private void setBackgroundColor(Scene scene) {
         scene.setFill(CityStyle.getBackgroundColor(styleProperty.getValue()));
+    }
+
+    public City createCity(CityNode resultNode, String cityName) {
+        // Add city to group and translate them according to model
+        int hoodHeight = 1;
+        int potentialHeight = resultNode.getSize() / 400;
+        if (potentialHeight > 1) {
+            hoodHeight = potentialHeight;
+        }
+        City city = new City(hoodHeight, resultNode.getSize(), resultNode.getSize(), cityName);
+        addAllNodesRecursively(city, resultNode.getChildren());
+
+        LayoutManager layoutManager = new LayoutManager();
+        layoutManager.moveCityChildrenBackToCity(city);
+        layoutManager.correctYDirection(city);
+        layoutManager.setRelativeOffset(city);
+        layoutManager.moveCityToChildren(city);
+
+        city.setTranslateX(city.getWidth() / 2);
+        city.setTranslateZ(city.getWidth() / 2);
+        city.setTranslateY(100);
+        return city;
+    }
+
+    private static final int HEIGHT_DIVISOR = 3;
+
+    private void addAllNodesRecursively(Hood hood, List<CityNode> children) {
+        for (CityNode node : children) {
+
+            String info = node.getModel().getInfo();
+            if (node.isLeaf()) {
+                double height = node.getModel().getLinesOfCode() / HEIGHT_DIVISOR;
+                hood.addBuilding(new Building(node.getSize(), height, node.getX(), node.getY(), info, node.getModel(), hood.getLevel() + 1)); // Y==Z
+            } else {
+                double incrementedHeight = hood.getHeight();
+                Hood childHood = new Hood(incrementedHeight, node.getSize(), node.getSize(), node.getX(), node.getY(), info, hood.getLevel() + 1);
+                hood.addHood(childHood);
+                addAllNodesRecursively(childHood, node.getChildren());
+            }
+        }
     }
 }
